+++ CHANGELOG.md
@@ -1,10 +1,14 @@
-alpha3:
+FUTURE PLANS:
     - [todo] night-vision google
     - [todo] enemies engine
     - [todo] actual enemies with some hunting logic
     - [todo] stealth movement
     - [todo] end game
-    
+
+alpha3:
+    - rewritten vegetation generation
+    - prepare for chunk world infinite rendering
+
 alpha2:
     - proper day/night cycle (40% day,10% evening, 40% night, 10% morning)
     - more code refactor, reduce of "magic numbers"